backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-jose
passlib[bcrypt]
asyncpg